backend/requirements.txt
# Core dependencies
fastapi>=0.95.1
uvicorn>=0.22.0
pydantic>=1.10.7
python-dotenv>=1.0.0
pyautogen==0.1.14
openai>=0.27.0

# Deep learning
torch>=2.0.1
torchvision>=0.13.0
torchaudio>=0.12.0
pytorch-lightning>=2.0.0
pytorch-forecasting>=1.0.0

# Data processing and analysis
pandas>=2.0.1
numpy>=1.24.3
scikit-learn>=1.2.2
scipy>=1.10.1
statsmodels>=0.14.0
prophet==1.1.4
matplotlib>=3.7.1
seaborn>=0.12.2
pyarrow>=12.0.1

# Time series analysis
darts==0.24.0

# Weather data
meteocalc==1.1.0
pyowm==3.3.0

# Database and storage
sqlalchemy>=2.0.13
alembic==1.10.4
pymongo>=4.3.3
redis>=4.5.4

# API and communication
httpx>=0.24.0
requests>=2.30.0
websockets==11.0.3
aiohttp>=3.8.4

# Logging and monitoring
loguru>=0.7.0
structlog==23.1.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.75.2

# Utilities
tqdm>=4.65.0
tenacity>=8.2.2
python-dateutil>=2.8.2
pytz>=2023.3
pyyaml==6.0

# Vector storage
faiss-cpu==1.7.4
hnswlib>=0.7.0
pinecone-client>=2.2.1

# Natural language processing
langchain>=0.0.166
tiktoken>=0.4.0

# Required dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pymongo>=4.3.3
psycopg[binary,pool]>=3.1.8
redis>=4.5.4
python-dotenv>=1.0.0
pydantic>=1.10.7
autogen>=0.2.0
pandas>=2.0.0
numpy>=1.24.2
scikit-learn>=1.2.2
matplotlib>=3.7.1
openai>=0.27.4
plotly>=5.14.1
statsmodels>=0.13.5
pytest>=7.3.1
httpx>=0.24.0
python-multipart>=0.0.6
email-validator>=2.0.0
pendulum>=2.1.2
pytz>=2023.3
loguru>=0.7.0

# Cài đặt thêm backoff để xử lý lỗi import
backoff>=2.2.1 